+++ tests/conftest.py
@@ -72,6 +72,7 @@ def minimal_attribute_dict():
             "columns": ["a", "c"],
             "new_column_names": "f",
             "pd_method_name": "sum",
+            "drop_original": True,
         },
         "BaseDateTransformer": {
             "columns": ["a"],
@@ -88,6 +89,7 @@ def minimal_attribute_dict():
         "EqualityChecker": {
             "columns": ["a", "b"],
             "new_col_name": "c",
+            "drop_original": True,
         },
         "DateDiffLeapYearTransformer": {
             "column_lower": "a",
@@ -198,6 +200,7 @@ def minimal_attribute_dict():
         },
         "OneHotEncodingTransformer": {
             "columns": ["c"],
+            "drop_original": True,
         },
         "LogTransformer": {
             "columns": ["a"],