+++ tubular/comparison.py
@@ -3,9 +3,10 @@ from __future__ import annotations
 import pandas as pd  # noqa: TCH002
 
 from tubular.base import BaseTwoColumnTransformer
+from tubular.mixins import BaseDropOriginalMixin
 
 
-class EqualityChecker(BaseTwoColumnTransformer):
+class EqualityChecker(BaseDropOriginalMixin, BaseTwoColumnTransformer):
     """Transformer to check if two columns are equal.
 
     Parameters
@@ -33,11 +34,7 @@ class EqualityChecker(BaseTwoColumnTransformer):
     ) -> None:
         super().__init__(columns=columns, new_col_name=new_col_name, **kwargs)
 
-        if not (isinstance(drop_original, bool)):
-            msg = f"{self.classname()}: drop_original should be bool"
-            raise TypeError(msg)
-
-        self.drop_original = drop_original
+        BaseDropOriginalMixin.set_drop_original_column(self, drop_original)
 
     def transform(self, X: pd.DataFrame) -> pd.DataFrame:
         """Create a column which is populated by the boolean
@@ -58,8 +55,12 @@ class EqualityChecker(BaseTwoColumnTransformer):
 
         X[self.new_col_name] = X[self.columns[0]] == X[self.columns[1]]
 
-        if self.drop_original:
-            for col in self.columns:
-                del X[col]
+        # Drop original columns if self.drop_original is True
+        BaseDropOriginalMixin.drop_original_column(
+            self,
+            X,
+            self.drop_original,
+            self.columns,
+        )
 
         return X