+++ CHANGELOG.rst
@@ -46,8 +46,10 @@ Moved tests for these objects to new approach.
 - Refactored ModeImputer tests in new format.
 - Added generic init tests to base tests for transformers that take two columns as an input.
 - Refactored EqualityChecker tests in new format.
-- Bugfix to MeanResponseTransformer to ignore unobserved categorical levels
+- Bugfix to MeanResponseTransformer to ignore unobserved categorical levels.
 - Added test_BaseTwoColumnTransformer base class for columns that require a list of two columns for input
+- Added BaseDropOriginalMixin to mixin transformers to handle validation and method of dropping original features, also added appropriate test classes.
+
 
 Removed
 ^^^^^^^